+++ fregot.cabal
@@ -44,6 +44,7 @@ Library
     Fregot.Eval.Number
     Fregot.Eval.TempObject
     Fregot.Eval.Value
+    Fregot.Eval.Value.Conversion
     Fregot.Find
     Fregot.Interpreter
     Fregot.Interpreter.Bundle
