+++ lib/Fregot/Builtins/Json.hs
@@ -32,7 +32,6 @@ builtins = HMS.fromList
 
 builtin_json_marshal :: Monad m => Builtin m
 builtin_json_marshal = Builtin
-    (In Out)
     (Ty.any 🡒 Ty.out Ty.string) $ pure $
     \(Cons val Nil) -> case Json.fromValue val of
         Left err   -> throwDoc err
@@ -40,7 +39,6 @@ builtin_json_marshal = Builtin
 
 builtin_json_unmarshal :: Monad m => Builtin m
 builtin_json_unmarshal = Builtin
-    (In Out)
     (Ty.string 🡒 Ty.out Ty.unknown) $ pure $
     \(Cons str Nil) -> case A.eitherDecodeStrict' (T.encodeUtf8 str) of
         Left  err -> throwString err
