+++ lib/Fregot/Types/Builtins.hs
@@ -13,18 +13,19 @@ Portability : POSIX
 {-# LANGUAGE Rank2Types     #-}
 {-# LANGUAGE TypeOperators  #-}
 module Fregot.Types.Builtins
-    ( InTypes (..)
+    ( TypeRepr (..)
     , BuiltinChecker (..)
-    , BuiltinType
+    , BuiltinType (..)
     , out
     , (🡒)
     ) where
 
+import           Fregot.Eval.Value.Conversion (FromVal, ToVal)
 import           Fregot.Types.Internal
 
-data InTypes (i :: [*]) where
-    Nil  :: InTypes '[]
-    Cons :: Type -> InTypes i -> InTypes (a ': i)
+data TypeRepr (i :: [*]) (o :: *) where
+    Out :: ToVal o   => Type -> TypeRepr '[] o
+    In  :: FromVal a => Type -> TypeRepr i o -> TypeRepr (a ': i) o
 
 data BuiltinChecker m = BuiltinChecker
     { bcUnify    :: Type -> Type -> m Type
@@ -32,12 +33,27 @@ data BuiltinChecker m = BuiltinChecker
     , bcCatch    :: forall a. m a -> m a -> m a
     }
 
-type BuiltinType (i :: [*]) =
-    forall m. Monad m => BuiltinChecker m -> InTypes i -> m Type
+-- | A builtin type has two representations.  One of them is a function that
+-- takes a number of types and returns the return type.  This allows us to do
+-- highly granular type checking.
+--
+-- The other representation is a deep embedding.  This can be used to "print"
+-- the type or render the capabilities doc.  It is also used to convert
+-- arguments to the right shape, so the arities must match.
+data BuiltinType (i :: [*]) (o :: *) = BuiltinType
+    { btCheck :: forall m. Monad m => BuiltinChecker m -> TypeRepr i o -> m Type
+    , btRepr  :: TypeRepr i o
+    }
 
-out :: Type -> BuiltinType '[]
-out t = \_ Nil -> pure t
+out :: ToVal o => Type -> BuiltinType '[] o
+out ty = BuiltinType
+    { btCheck = \_ _ -> pure ty
+    , btRepr  = Out ty
+    }
 
-(🡒) :: Type -> BuiltinType i -> BuiltinType (a ': i)
-(🡒) expect f = \c (Cons actual t) -> bcSubsetOf c actual expect >> f c t
+(🡒) :: FromVal a => Type -> BuiltinType i o -> BuiltinType (a ': i) o
+(🡒) expect bt = BuiltinType
+    { btCheck = \c (In actual t) -> bcSubsetOf c actual expect >> btCheck bt c t
+    , btRepr  = In expect (btRepr bt)
+    }
 infixr 6 🡒
